+++ R/server.R
@@ -1,7 +1,4 @@
-# I thought these were going to be modules, but it's just a pair of functions
-# that you can call from other things, I think.
-
-#' Create or Update a Cookie
+#' Create or update a cookie
 #'
 #' Instruct the user's browser to create a cookie via JavaScript.
 #'
@@ -9,18 +6,15 @@
 #' @inheritParams shiny::moduleServer
 #'
 #' @export
-set_cookie <- function(name,
-                       contents,
+set_cookie <- function(cookie_name,
+                       cookie_value,
                        expiration = 90,
                        secure_only = NULL,
                        domain = NULL,
                        path = NULL,
                        same_site = NULL,
                        session = shiny::getDefaultReactiveDomain()) {
-  # This sets via javascript. When we add HttpOnly handling, this should either
-  # deal with that or error if they try to do http_only.
-
-  attributes <- .validate_attributes(
+  attributes <- .javascript_attributes(
     expiration = expiration,
     secure_only = secure_only,
     domain = domain,
@@ -31,14 +25,14 @@ set_cookie <- function(name,
   session$sendCustomMessage(
     "cookie-set",
     list(
-      name = name,
-      value = contents,
+      name = cookie_name,
+      value = cookie_value,
       attributes = attributes
     )
   )
 }
 
-#' Remove a Cookie
+#' Remove a cookie
 #'
 #' Instruct the user's browser to remove a cookie via JavaScript.
 #'
@@ -46,10 +40,10 @@ set_cookie <- function(name,
 #' @inheritParams shiny::moduleServer
 #'
 #' @export
-remove_cookie <- function(name,
+remove_cookie <- function(cookie_name,
                           session = shiny::getDefaultReactiveDomain()) {
   session$sendCustomMessage(
     "cookie-remove",
-    list(name = name)
+    list(name = cookie_name)
   )
 }